Move quick-start extensions to admin.connections feature
