+++ features/admin.connections.v1/components/authenticators/email-otp/assets/build-login-flow.png
