+++ features/admin.connections.v1/components/authenticators/fido/assets/build-login-flow.png
