+++ features/admin.connections.v1/components/authenticators/fido/assets/configure-parameters.png
