+++ features/admin.connections.v1/components/authenticators/magic-link/assets/build-login-flow.png
