+++ features/admin.connections.v1/components/authenticators/sms-otp/assets/build-login-flow.png
