+++ features/admin.connections.v1/components/authenticators/sms-otp/quick-start.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021-2024, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2022-2024, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,37 +16,40 @@
  * under the License.
  */
 
+import {
+    VerticalStepper,
+    VerticalStepperStepInterface
+} from "@wso2is/admin.core.v1/components/vertical-stepper/vertical-stepper";
 import { FeatureConfigInterface } from "@wso2is/admin.core.v1/models";
 import { AppState } from "@wso2is/admin.core.v1/store";
+import ApplicationSelectionModal from "@wso2is/admin.extensions.v1/components/shared/application-selection-modal";
 import { hasRequiredScopes } from "@wso2is/core/helpers";
-import { TestableComponentInterface } from "@wso2is/core/models";
+import { IdentifiableComponentInterface } from "@wso2is/core/models";
 import { GenericIcon, Heading, Link, PageHeader, Text } from "@wso2is/react-components";
 import React, { FunctionComponent, ReactElement, useMemo, useState } from "react";
 import { Trans, useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import { Grid } from "semantic-ui-react";
 import BuildLoginFlowIllustration from "./assets/build-login-flow.png";
-import { VerticalStepper, VerticalStepperStepInterface } from "../../component-extensions";
-import ApplicationSelectionModal from "../../shared/application-selection-modal";
 
 /**
  * Prop types of the component.
  */
-type TOTPQuickStartPropsInterface = TestableComponentInterface;
+type SMSOTPQuickStartPropsInterface = IdentifiableComponentInterface;
 
 /**
- * Quick start content for the TOTP authenticator.
+ * Quick start content for the SMS OTP authenticator.
  *
  * @param props - Props injected into the component.
  *
- * @returns TOTP authenticator quick start component.
+ * @returns SMS OTP Quick start component.
  */
-const TOTPQuickStart: FunctionComponent<TOTPQuickStartPropsInterface> = (
-    props: TOTPQuickStartPropsInterface
+const SMSOTPQuickStart: FunctionComponent<SMSOTPQuickStartPropsInterface> = (
+    props: SMSOTPQuickStartPropsInterface
 ): ReactElement => {
 
     const {
-        ["data-testid"]: testId
+        [ "data-componentid" ]: componentId
     } = props;
 
     const { t } = useTranslation();
@@ -63,7 +66,7 @@ const TOTPQuickStart: FunctionComponent<TOTPQuickStartPropsInterface> = (
 
     /**
      * Vertical Stepper steps.
-     * @returns List of steps.
+     * @returns An array of Vertical Stepper steps.
      */
     const steps: VerticalStepperStepInterface[] = [
         {
@@ -72,36 +75,37 @@ const TOTPQuickStart: FunctionComponent<TOTPQuickStartPropsInterface> = (
                     <Text>
                         <Trans
                             i18nKey={
-                                "extensions:develop.identityProviders.totp.quickStart.steps.selectApplication.content"
+                                "extensions:develop.identityProviders.smsOTP.quickStart.steps.selectApplication.content"
                             }
                         >
                             Choose the { isApplicationReadAccessAllowed ? (
                                 <Link external={ false } onClick={ () => setShowApplicationModal(true) }>
                                 application </Link>) : "application" }
-                            for which you want to set up TOTP login.
+                            for which you want to set up SMS OTP login.
                         </Trans>
                     </Text>
                 </>
             ),
-            stepTitle: t("extensions:develop.identityProviders.totp.quickStart.steps.selectApplication.heading")
+            stepTitle: t("extensions:develop.identityProviders.smsOTP.quickStart.steps.selectApplication.heading")
         },
         {
             stepContent: (
                 <>
                     <Text>
                         <Trans
-                            i18nKey={ "extensions:develop.identityProviders.totp.quickStart.steps.selectTOTP.content" }
+                            i18nKey={ "extensions:develop.identityProviders.smsOTP.quickStart.steps.selectSMSOTP" +
+                                ".content" }
                         >
-                            Go to <strong>Login Flow</strong> tab and click on the <strong>Username & Password + TOTP
-                            </strong> option from the Multi-factor login section to configure a basic TOTP flow.
+                            Go to <strong>Login Flow</strong> tab and click on the <strong>Username & Password + SMS OTP
+                            </strong> option from the Multi-factor login section to configure a basic SMS OTP flow.
                         </Trans>
                     </Text>
                     <GenericIcon inline transparent icon={ BuildLoginFlowIllustration } size="huge"/>
                 </>
             ),
             stepTitle: (
-                <Trans i18nKey="extensions:develop.identityProviders.totp.quickStart.steps.selectTOTP.heading">
-                    Select <strong>TOTP</strong> option
+                <Trans i18nKey="extensions:develop.identityProviders.smsOTP.quickStart.steps.selectSMSOTP.heading">
+                    Select <strong>SMS OTP</strong> option
                 </Trans>
             )
         }
@@ -109,17 +113,17 @@ const TOTPQuickStart: FunctionComponent<TOTPQuickStartPropsInterface> = (
 
     return (
         <>
-            <Grid data-testid={ testId } className="authenticator-quickstart-content">
+            <Grid data-testid={ componentId } className="authenticator-quickstart-content">
                 <Grid.Row textAlign="left">
                     <Grid.Column width={ 16 }>
                         <PageHeader
                             className="mb-2"
+                            title={ t("extensions:develop.identityProviders.smsOTP.quickStart.heading") }
                             imageSpaced={ false }
                             bottomMargin={ false }
-                            title={ t("extensions:develop.identityProviders.totp.quickStart.heading") }
                         />
                         <Heading subHeading as="h6">
-                            { t("extensions:develop.identityProviders.totp.quickStart.subHeading") }
+                            { t("extensions:develop.identityProviders.smsOTP.quickStart.subHeading") }
                         </Heading>
                     </Grid.Column>
                 </Grid.Row>
@@ -137,12 +141,14 @@ const TOTPQuickStart: FunctionComponent<TOTPQuickStartPropsInterface> = (
             {
                 showApplicationModal && (
                     <ApplicationSelectionModal
-                        data-testid={ `${ testId }-application-selection-modal` }
+                        data-testid={ `${ componentId }-application-selection-modal` }
                         open={ showApplicationModal }
                         onClose={ () => setShowApplicationModal(false) }
-                        heading={ t("extensions:develop.identityProviders.totp.quickStart.addLoginModal.heading") }
+                        heading={
+                            t("extensions:develop.identityProviders.smsOTP.quickStart.addLoginModal.heading")
+                        }
                         subHeading={
-                            t("extensions:develop.identityProviders.totp.quickStart.addLoginModal.subHeading")
+                            t("extensions:develop.identityProviders.smsOTP.quickStart.addLoginModal.subHeading")
                         }
                         data-componentid="connections"
                     />
@@ -155,8 +161,8 @@ const TOTPQuickStart: FunctionComponent<TOTPQuickStartPropsInterface> = (
 /**
  * Default props for the component
  */
-TOTPQuickStart.defaultProps = {
-    "data-testid": "totp-authenticator-quick-start"
+SMSOTPQuickStart.defaultProps = {
+    "data-componentid": "sms-otp-authenticator-quick-start"
 };
 
 /**
@@ -164,4 +170,4 @@ TOTPQuickStart.defaultProps = {
  * TODO: Change this to a named export once react starts supporting named exports for code splitting.
  * @see {@link https://reactjs.org/docs/code-splitting.html#reactlazy}
  */
-export default TOTPQuickStart;
+export default SMSOTPQuickStart;
