+++ features/admin.connections.v1/components/authenticators/totp/quick-start.tsx
@@ -16,8 +16,13 @@
  * under the License.
  */
 
-import { AppState } from "@wso2is/admin.core.v1";
+import {
+    VerticalStepper,
+    VerticalStepperStepInterface
+} from "@wso2is/admin.core.v1/components/vertical-stepper/vertical-stepper";
 import { FeatureConfigInterface } from "@wso2is/admin.core.v1/models";
+import { AppState } from "@wso2is/admin.core.v1/store";
+import ApplicationSelectionModal from "@wso2is/admin.extensions.v1/components/shared/application-selection-modal";
 import { hasRequiredScopes } from "@wso2is/core/helpers";
 import { TestableComponentInterface } from "@wso2is/core/models";
 import { GenericIcon, Heading, Link, PageHeader, Text } from "@wso2is/react-components";
@@ -26,27 +31,25 @@ import { Trans, useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import { Grid } from "semantic-ui-react";
 import BuildLoginFlowIllustration from "./assets/build-login-flow.png";
-import { VerticalStepper, VerticalStepperStepInterface } from "../../component-extensions";
-import ApplicationSelectionModal from "../../shared/application-selection-modal";
 
 /**
  * Prop types of the component.
  */
-type EmailOTPQuickStartPropsInterface = TestableComponentInterface;
+type TOTPQuickStartPropsInterface = TestableComponentInterface;
 
 /**
- * Quick start content for the Email OTP authenticator.
+ * Quick start content for the TOTP authenticator.
  *
  * @param props - Props injected into the component.
  *
- * @returns email OTP authenticator quick start component.
+ * @returns TOTP authenticator quick start component.
  */
-const EmailOTPQuickStart: FunctionComponent<EmailOTPQuickStartPropsInterface> = (
-    props: EmailOTPQuickStartPropsInterface
+const TOTPQuickStart: FunctionComponent<TOTPQuickStartPropsInterface> = (
+    props: TOTPQuickStartPropsInterface
 ): ReactElement => {
 
     const {
-        [ "data-testid" ]: testId
+        ["data-testid"]: testId
     } = props;
 
     const { t } = useTranslation();
@@ -72,39 +75,36 @@ const EmailOTPQuickStart: FunctionComponent<EmailOTPQuickStartPropsInterface> = 
                     <Text>
                         <Trans
                             i18nKey={
-                                "extensions:develop.identityProviders.emailOTP.quickStart.steps.selectApplication" +
-                                ".content"
+                                "extensions:develop.identityProviders.totp.quickStart.steps.selectApplication.content"
                             }
                         >
                             Choose the { isApplicationReadAccessAllowed ? (
                                 <Link external={ false } onClick={ () => setShowApplicationModal(true) }>
                                 application </Link>) : "application" }
-                            for which you want to set up Email OTP login.
+                            for which you want to set up TOTP login.
                         </Trans>
                     </Text>
                 </>
             ),
-            stepTitle: t("extensions:develop.identityProviders.emailOTP.quickStart.steps.selectApplication.heading")
+            stepTitle: t("extensions:develop.identityProviders.totp.quickStart.steps.selectApplication.heading")
         },
         {
             stepContent: (
                 <>
                     <Text>
                         <Trans
-                            i18nKey={ "extensions:develop.identityProviders.emailOTP.quickStart.steps.selectEmailOTP" +
-                            ".content" }
+                            i18nKey={ "extensions:develop.identityProviders.totp.quickStart.steps.selectTOTP.content" }
                         >
-                            Go to <strong>Login Flow</strong> tab and click on the <strong>Username & Password +
-                                Email OTP </strong> option from the Multi-factor login section to configure a
-                                basic Email OTP flow.
+                            Go to <strong>Login Flow</strong> tab and click on the <strong>Username & Password + TOTP
+                            </strong> option from the Multi-factor login section to configure a basic TOTP flow.
                         </Trans>
                     </Text>
                     <GenericIcon inline transparent icon={ BuildLoginFlowIllustration } size="huge"/>
                 </>
             ),
             stepTitle: (
-                <Trans i18nKey="extensions:develop.identityProviders.emailOTP.quickStart.steps.selectEmailOTP.heading">
-                    Select <strong>Email OTP</strong> option
+                <Trans i18nKey="extensions:develop.identityProviders.totp.quickStart.steps.selectTOTP.heading">
+                    Select <strong>TOTP</strong> option
                 </Trans>
             )
         }
@@ -117,12 +117,12 @@ const EmailOTPQuickStart: FunctionComponent<EmailOTPQuickStartPropsInterface> = 
                     <Grid.Column width={ 16 }>
                         <PageHeader
                             className="mb-2"
-                            title={ t("extensions:develop.identityProviders.emailOTP.quickStart.heading") }
                             imageSpaced={ false }
                             bottomMargin={ false }
+                            title={ t("extensions:develop.identityProviders.totp.quickStart.heading") }
                         />
                         <Heading subHeading as="h6">
-                            { t("extensions:develop.identityProviders.emailOTP.quickStart.subHeading") }
+                            { t("extensions:develop.identityProviders.totp.quickStart.subHeading") }
                         </Heading>
                     </Grid.Column>
                 </Grid.Row>
@@ -143,11 +143,9 @@ const EmailOTPQuickStart: FunctionComponent<EmailOTPQuickStartPropsInterface> = 
                         data-testid={ `${ testId }-application-selection-modal` }
                         open={ showApplicationModal }
                         onClose={ () => setShowApplicationModal(false) }
-                        heading={
-                            t("extensions:develop.identityProviders.emailOTP.quickStart.addLoginModal.heading")
-                        }
+                        heading={ t("extensions:develop.identityProviders.totp.quickStart.addLoginModal.heading") }
                         subHeading={
-                            t("extensions:develop.identityProviders.emailOTP.quickStart.addLoginModal.subHeading")
+                            t("extensions:develop.identityProviders.totp.quickStart.addLoginModal.subHeading")
                         }
                         data-componentid="connections"
                     />
@@ -160,8 +158,8 @@ const EmailOTPQuickStart: FunctionComponent<EmailOTPQuickStartPropsInterface> = 
 /**
  * Default props for the component
  */
-EmailOTPQuickStart.defaultProps = {
-    "data-testid": "email-otp-authenticator-quick-start"
+TOTPQuickStart.defaultProps = {
+    "data-testid": "totp-authenticator-quick-start"
 };
 
 /**
@@ -169,4 +167,4 @@ EmailOTPQuickStart.defaultProps = {
  * TODO: Change this to a named export once react starts supporting named exports for code splitting.
  * @see {@link https://reactjs.org/docs/code-splitting.html#reactlazy}
  */
-export default EmailOTPQuickStart;
+export default TOTPQuickStart;
