+++ features/admin.extensions.v1/configs/identity-provider.tsx
@@ -17,15 +17,15 @@
  */
 
 import { ConnectionManagementConstants, ConnectionTabTypes } from "@wso2is/admin.connections.v1";
+import {
+    SmsOTPAuthenticator
+} from "@wso2is/admin.connections.v1/components/authenticators/sms-otp/sms-otp-authenticator";
 import { IdentityProviderManagementConstants } from "@wso2is/admin.identity-providers.v1/constants";
 import { IdentityProviderTabTypes } from "@wso2is/admin.identity-providers.v1/models";
 import { I18n } from "@wso2is/i18n";
 import { ResourceTabPaneInterface } from "@wso2is/react-components";
 import React, { ElementType, ReactElement } from "react";
 import { IdentityProviderConfig } from "./models";
-import {
-    SmsOTPAuthenticator
-} from "../components/authenticators/sms-otp/sms-otp-authenticator";
 import QuickStartTab from "../components/component-extensions/application/quick-start-tab";
 import { SIWEAuthenticatorForm } from "../identity-provider-templates/templates/swe/swe-authenticator-form";
 import SIWEIdPTemplate from "../identity-provider-templates/templates/swe/swe.json";
